requirements.txt
fastapi
tortoise-orm[asyncpg]
uvicorn
httpx
pydantic-settings
aerich